# docs/swagger.yaml
basePath: /api
definitions:
  models.AuthResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  models.CreateExpenseRequest:
    properties:
      amount:
        type: number
      category:
        type: string
      currency:
        type: string
      description:
        type: string
      expense_date:
        type: string
      receipt_image_url:
        type: string
      team_id:
        type: string
    required:
    - amount
    - category
    - currency
    - description
    - expense_date
    type: object
  models.ErrorResponse:
    properties:
      message:
        example: invalid request
        type: string
    type: object
  models.Expense:
    properties:
      amount:
        type: number
      category:
        type: string
      created_at:
        type: string
      currency:
        type: string
      description:
        type: string
      expense_date:
        description: YYYY-MM-DD
        type: string
      id:
        type: string
      receipt_image_url:
        type: string
      status:
        description: pending, approved, rejected
        type: string
      team_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  models.RegisterRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    type: object
  models.UpdateExpenseRequest:
    properties:
      amount:
        type: number
      category:
        type: string
      description:
        type: string
      expense_date:
        type: string
      status:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:909
info:
  contact:
    email: mahingarodin@gmail.com
    name: pocketPilotBe-Rodin
  description: Backend API for PocketPilot expense tracking app.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: PocketPilot API
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT
      parameters:
      - description: Login payload
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Login user
      tags:
      - Auth
  /api/auth/profile:
    get:
      description: Retrieve authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - Auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: Register payload
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Register a new user
      tags:
      - Auth
  /api/expenses:
    get:
      description: Retrieve user's expenses
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Expense'
            type: array
      security:
      - BearerAuth: []
      summary: Get expenses
      tags:
      - Expenses
    post:
      consumes:
      - application/json
      description: Create a new expense
      parameters:
      - description: Expense payload
        in: body
        name: expense
        required: true
        schema:
          $ref: '#/definitions/models.CreateExpenseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Expense'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create expense
      tags:
      - Expenses
  /api/expenses/{id}:
    delete:
      description: Delete an expense
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete expense
      tags:
      - Expenses
    get:
      description: Retrieve a specific expense by ID
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Expense'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get expense
      tags:
      - Expenses
    put:
      consumes:
      - application/json
      description: Update an existing expense
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: string
      - description: Expense payload
        in: body
        name: expense
        required: true
        schema:
          $ref: '#/definitions/models.UpdateExpenseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Expense'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update expense
      tags:
      - Expenses
  /api/expenses/team/{teamId}:
    get:
      description: Retrieve expenses for a team
      parameters:
      - description: Team ID
        in: path
        name: teamId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Expense'
            type: array
      security:
      - BearerAuth: []
      summary: Get team expenses
      tags:
      - Expenses
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"